refactor: 게시물 상세보기 경로를 /recruitment-boards/[id]에서 /recruitment-boards/detail?id=[id]로 변경
