+++ src/app/components/recruitment-boards/detail/ApplyButton.tsx
