+++ src/app/components/recruitment-boards/detail/CheckApplicantButton.tsx
