+++ src/app/components/recruitment-boards/detail/RecruitmentBoardContent.tsx
