+++ src/app/components/recruitment-boards/detail/recruitmentBoardContent.module.scss
