+++ src/app/components/recruitment-boards/detail/recruitmentBoardTitle.module.scss
