+++ src/app/recruitment-boards/apply/page.tsx
@@ -1,16 +1,19 @@
 import ApplyBanner from '@/app/components/recruitment-boards/apply/ApplyBanner';
 import styles from './page.module.scss'
-import BoardDetail from '@/app/components/recruitment-boards/apply/BoardDetail';
 import ApplyForm from '@/app/components/recruitment-boards/apply/ApplyForm';
-import { RecruitmentBoardTitle } from '@/app/components/recruitment-boards/[id]/RecruitmentBoardTitle';
+import {
+  RecruitmentBoardDetailProvider
+} from '@/app/components/recruitment-boards/detail/RecruitmentBoardDetailProvider';
+import RecruitmentApplayContainer from '@/app/components/recruitment-boards/apply/RecruitmentApplayContainer';
 
 export default function Page() {
   return (
     <main className={styles.block}>
-      <ApplyBanner />
-      <RecruitmentBoardTitle title={'제목1'} type={'STUDY'} tag={'FRONT'} />
-      <BoardDetail/>
-      <ApplyForm/>
+      <RecruitmentBoardDetailProvider>
+        <ApplyBanner/>
+          <RecruitmentApplayContainer />
+        <ApplyForm/>
+      </RecruitmentBoardDetailProvider>
     </main>
   )
 }
