+++ src/app/recruitment-boards/detail/page.module.scss
