+++ src/app/recruitment-boards/list/page.tsx
@@ -1,4 +1,4 @@
-import { RecruitmentBoardTitle } from '@/app/components/recruitment-boards/[id]/RecruitmentBoardTitle';
+import { RecruitmentBoardTitle } from '@/app/components/recruitment-boards/detail/RecruitmentBoardTitle';
 import styles from './page.module.scss';
 import ListTable from '@/app/components/recruitment-boards/list/ListTable';
 import { Suspense } from 'react';
